+++ src/idangerous.swiper.js
@@ -4,7 +4,7 @@ var Swiper = function (selector, params) {
     /*=========================
       A little bit dirty but required part for IE8 and old FF support
       ===========================*/
-    if (document.body.__defineGetter__) {
+    if (!document.body.outerHTML && document.body.__defineGetter__) {
         if (HTMLElement) {
             var element = HTMLElement.prototype;
             if (element.__defineGetter__) {
@@ -160,7 +160,8 @@ var Swiper = function (selector, params) {
         roundLengths: false,
         //Auto Height
         calculateHeight: false,
-        cssWidthAndHeight: false,
+        //Apply CSS for width and/or height
+        cssWidthAndHeight: false, // or true or 'width' or 'height'
         //Images Preloader
         updateOnImagesReady : true,
         //Form elements
@@ -191,6 +192,9 @@ var Swiper = function (selector, params) {
             logic: 'reload', //or 'change'
             loadAllSlides: false
         },
+        // One way swipes
+        swipeToPrev: true,
+        swipeToNext: true,
         //Namespace
         slideElement: 'div',
         slideClass: 'swiper-slide',
@@ -553,7 +557,7 @@ var Swiper = function (selector, params) {
                 }
             }
         } else if (Object.prototype.toString.call(callback) === '[object String]') {
-            if (params['on' + callback]) _this.fireCallback(params['on' + callback]);
+            if (params['on' + callback]) _this.fireCallback(params['on' + callback], arguments[1], arguments[2], arguments[3], arguments[4], arguments[5]);
         } else {
             callback(arguments[1], arguments[2], arguments[3], arguments[4], arguments[5]);
         }
@@ -628,7 +632,7 @@ var Swiper = function (selector, params) {
         var _width = _this.h.getWidth(_this.container, false, params.roundLengths);
         var _height = _this.h.getHeight(_this.container, false, params.roundLengths);
         if (_width === _this.width && _height === _this.height && !force) return;
-        
+
         _this.width = _width;
         _this.height = _height;
 
@@ -735,7 +739,7 @@ var Swiper = function (selector, params) {
                                 _this.snapGrid.push(slideLeft);
                             }
                         }
-                            
+
                     }
                     else {
                         _this.snapGrid.push(slideLeft);
@@ -799,11 +803,11 @@ var Swiper = function (selector, params) {
             }
             else {
                 slideHeight = isH ? _this.height : _this.height / params.slidesPerView;
-                if (params.roundLengths) slideHeight = Math.round(slideHeight);
+                if (params.roundLengths) slideHeight = Math.ceil(slideHeight);
                 wrapperHeight = isH ? _this.height : _this.slides.length * slideHeight;
             }
             slideWidth = isH ? _this.width / params.slidesPerView : _this.width;
-            if (params.roundLengths) slideWidth = Math.round(slideWidth);
+            if (params.roundLengths) slideWidth = Math.ceil(slideWidth);
             wrapperWidth = isH ? _this.slides.length * slideWidth : _this.width;
             slideSize = isH ? slideWidth : slideHeight;
 
@@ -843,13 +847,11 @@ var Swiper = function (selector, params) {
             }
 
             wrapperSize = isH ? wrapperWidth + _this.wrapperRight + _this.wrapperLeft : wrapperHeight + _this.wrapperTop + _this.wrapperBottom;
-            if (!params.cssWidthAndHeight) {
-                if (parseFloat(wrapperWidth) > 0) {
-                    wrapper.style.width = wrapperWidth + 'px';
-                }
-                if (parseFloat(wrapperHeight) > 0) {
-                    wrapper.style.height = wrapperHeight + 'px';
-                }
+            if (parseFloat(wrapperWidth) > 0 && (!params.cssWidthAndHeight || params.cssWidthAndHeight === 'height')) {
+                wrapper.style.width = wrapperWidth + 'px';
+            }
+            if (parseFloat(wrapperHeight) > 0 && (!params.cssWidthAndHeight || params.cssWidthAndHeight === 'width')) {
+                wrapper.style.height = wrapperHeight + 'px';
             }
             slideLeft = 0;
             _this.snapGrid = [];
@@ -858,13 +860,11 @@ var Swiper = function (selector, params) {
                 _this.snapGrid.push(slideLeft);
                 _this.slidesGrid.push(slideLeft);
                 slideLeft += slideSize;
-                if (!params.cssWidthAndHeight) {
-                    if (parseFloat(slideWidth) > 0) {
-                        _this.slides[i].style.width = slideWidth + 'px';
-                    }
-                    if (parseFloat(slideHeight) > 0) {
-                        _this.slides[i].style.height = slideHeight + 'px';
-                    }
+                if (parseFloat(slideWidth) > 0 && (!params.cssWidthAndHeight || params.cssWidthAndHeight === 'height')) {
+                    _this.slides[i].style.width = slideWidth + 'px';
+                }
+                if (parseFloat(slideHeight) > 0 && (!params.cssWidthAndHeight || params.cssWidthAndHeight === 'width')) {
+                    _this.slides[i].style.height = slideHeight + 'px';
                 }
             }
 
@@ -992,7 +992,8 @@ var Swiper = function (selector, params) {
         function _loadImage(src) {
             var image = new Image();
             image.onload = function () {
-                if (_this && _this.imagesLoaded !== undefined) _this.imagesLoaded++;
+                if (typeof _this === 'undefined' || _this === null) return;
+                if (_this.imagesLoaded !== undefined) _this.imagesLoaded++;
                 if (_this.imagesLoaded === _this.imagesToLoad.length) {
                     _this.reInit();
                     if (params.onImagesReady) _this.fireCallback(params.onImagesReady, _this);
@@ -1270,6 +1271,19 @@ var Swiper = function (selector, params) {
         else e.returnValue = false;
         return false;
     }
+    _this.disableMousewheelControl = function () {
+        if (!_this._wheelEvent) return false;
+        params.mousewheelControl = false;
+        _this.h.removeEventListener(_this.container, _this._wheelEvent, handleMousewheel);
+        return true;
+    };
+
+    _this.enableMousewheelControl = function () {
+        if (!_this._wheelEvent) return false;
+        params.mousewheelControl = true;
+        _this.h.addEventListener(_this.container, _this._wheelEvent, handleMousewheel);
+        return true;
+    };
 
     /*=========================
       Grab Cursor
@@ -1351,8 +1365,18 @@ var Swiper = function (selector, params) {
             return false;
         }
 
-        if (params.noSwiping && (event.target || event.srcElement) && noSwipingSlide(event.target || event.srcElement)) return false;
-        if (params.ignoreTouchClass && (event.target || event.srcElement) && isIgnoreElement(event.target || event.srcElement)) {
+        // Blur active elements
+        var eventTarget = event.target || event.srcElement;
+        if (document.activeElement) {
+            if (document.activeElement !== eventTarget) document.activeElement.blur();
+        }
+
+        // Form tag names
+        var formTagNames = ('input select textarea').split(' ');
+
+        // Check for no swiping
+        if (params.noSwiping && eventTarget && noSwipingSlide(eventTarget)) return false;
+        if (params.ignoreTouchClass && eventTarget && isIgnoreElement(eventTarget)) {
             _this.isTouched = false;
             return;
         }
@@ -1362,10 +1386,13 @@ var Swiper = function (selector, params) {
         _this.receivedMove = false;
         isTouchEvent = event.type === 'touchstart';
 
+        // prevent user enter with right and the swiper move (needs isTouchEvent)
+        if (!isTouchEvent && 'which' in event && event.which === 3) return false;
+
         if (!isTouchEvent || event.targetTouches.length === 1) {
             _this.callPlugins('onTouchStartBegin');
+            if (!isTouchEvent && !_this.isAndroid && formTagNames.indexOf(eventTarget.tagName.toLowerCase()) < 0) {
 
-            if (!isTouchEvent && !_this.isAndroid) {
                 if (event.preventDefault) event.preventDefault();
                 else event.returnValue = false;
             }
@@ -1431,6 +1458,18 @@ var Swiper = function (selector, params) {
         if (isScrolling) {
             _this.isTouched = false;
             return;
+        }
+
+        // One way swipes
+        if (isH) {
+            if ((!params.swipeToNext && pageX < _this.touches.startX) || ((!params.swipeToPrev && pageX > _this.touches.startX))) {
+                return;
+            }
+        }
+        else {
+            if ((!params.swipeToNext && pageY < _this.touches.startY) || ((!params.swipeToPrev && pageY > _this.touches.startY))) {
+                return;
+            }
         }
 
         //Check For Nested Swipers
@@ -1641,6 +1680,7 @@ var Swiper = function (selector, params) {
 
         setTimeout(function () {
             //Release inner links
+            if (typeof _this === 'undefined' || _this === null) return;
             if (params.preventLinks) {
                 _this.allowLinks = true;
             }
@@ -1984,6 +2024,7 @@ var Swiper = function (selector, params) {
         }
         else {
             newPosition = currentPosition < 0 ? Math.round(currentPosition / groupSize) * groupSize : 0;
+            if (currentPosition <= -maxWrapperPosition()) newPosition = -maxWrapperPosition();
         }
         if (params.scrollContainer)  {
             newPosition = currentPosition < 0 ? currentPosition : 0;
@@ -2041,7 +2082,7 @@ var Swiper = function (selector, params) {
             currentPosition += animationStep * time / (1000 / 60);
             condition = direction === 'toNext' ? currentPosition > newPosition : currentPosition < newPosition;
             if (condition) {
-                _this.setWrapperTranslate(Math.round(currentPosition));
+                _this.setWrapperTranslate(Math.ceil(currentPosition));
                 _this._DOMAnimating = true;
                 window.setTimeout(function () {
                     anim();
@@ -2050,12 +2091,12 @@ var Swiper = function (selector, params) {
             else {
                 if (params.onSlideChangeEnd) {
                     if (action === 'to') {
-                        if (toOptions.runCallbacks === true) _this.fireCallback(params.onSlideChangeEnd, _this);
+                        if (toOptions.runCallbacks === true) _this.fireCallback(params.onSlideChangeEnd, _this, direction);
                     }
                     else {
-                        _this.fireCallback(params.onSlideChangeEnd, _this);
+                        _this.fireCallback(params.onSlideChangeEnd, _this, direction);
                     }
-                    
+
                 }
                 if (params.onReachEnd && _this.activeIndex === _this.slides.length - 1) {
                     if (action === 'to') {
@@ -2344,6 +2385,7 @@ var Swiper = function (selector, params) {
         for (var i = 0; i < pagers.length; i++) {
             if (target === pagers[i]) index = i;
         }
+        if (params.autoplay) _this.stopAutoplay(true);
         _this.swipeTo(index);
     }
     _this.updatePagination = function (position) {
@@ -2723,7 +2765,8 @@ Swiper.prototype = {
             events = ['webkitTransitionEnd', 'transitionend', 'oTransitionEnd', 'MSTransitionEnd', 'msTransitionEnd'],
             i;
 
-        function fireCallBack() {
+        function fireCallBack(e) {
+            if (e.target !== el) return;
             callback(a);
             if (a.params.queueEndCallbacks) a._queueEndCallbacks = false;
             if (!permanent) {
@@ -2843,11 +2886,11 @@ Swiper.prototype = {
             var width = window.getComputedStyle(el, null).getPropertyValue('width');
             var returnWidth = parseFloat(width);
             //IE Fixes
-            if (isNaN(returnWidth) || width.indexOf('%') > 0) {
+            if (isNaN(returnWidth) || width.indexOf('%') > 0 || returnWidth < 0) {
                 returnWidth = el.offsetWidth - parseFloat(window.getComputedStyle(el, null).getPropertyValue('padding-left')) - parseFloat(window.getComputedStyle(el, null).getPropertyValue('padding-right'));
             }
             if (outer) returnWidth += parseFloat(window.getComputedStyle(el, null).getPropertyValue('padding-left')) + parseFloat(window.getComputedStyle(el, null).getPropertyValue('padding-right'));
-            if (round) return Math.round(returnWidth);
+            if (round) return Math.ceil(returnWidth);
             else return returnWidth;
         },
         getHeight: function (el, outer, round) {
@@ -2857,11 +2900,11 @@ Swiper.prototype = {
             var height = window.getComputedStyle(el, null).getPropertyValue('height');
             var returnHeight = parseFloat(height);
             //IE Fixes
-            if (isNaN(returnHeight) || height.indexOf('%') > 0) {
+            if (isNaN(returnHeight) || height.indexOf('%') > 0 || returnHeight < 0) {
                 returnHeight = el.offsetHeight - parseFloat(window.getComputedStyle(el, null).getPropertyValue('padding-top')) - parseFloat(window.getComputedStyle(el, null).getPropertyValue('padding-bottom'));
             }
             if (outer) returnHeight += parseFloat(window.getComputedStyle(el, null).getPropertyValue('padding-top')) + parseFloat(window.getComputedStyle(el, null).getPropertyValue('padding-bottom'));
-            if (round) return Math.round(returnHeight);
+            if (round) return Math.ceil(returnHeight);
             else return returnHeight;
         },
         getOffset: function (el) {
@@ -2991,7 +3034,7 @@ Swiper.prototype = {
 
         classList : (function () {
             'use strict';
-            var div = document.createElement('div').style;
+            var div = document.createElement('div');
             return 'classList' in div;
         })()
     },
@@ -3022,9 +3065,16 @@ if (window.jQuery || window.Zepto) {
     (function ($) {
         'use strict';
         $.fn.swiper = function (params) {
-            var s = new Swiper($(this)[0], params);
-            $(this).data('swiper', s);
-            return s;
+            var firstInstance;
+            this.each(function (i) {
+                var that = $(this);
+                if (!that.data('swiper')) {
+                    var s = new Swiper(that[0], params);
+                    if (!i) firstInstance = s;
+                    that.data('swiper', s);
+                }
+            });
+            return firstInstance;
         };
     })(window.jQuery || window.Zepto);
 }