+++ CHANGELOG.md
@@ -1,4 +1,24 @@
 # Change Log
+## Swiper 2.7.0 - Updated on August 30, 2014
+
+  * New API parameters "swipeToNext", "swipeToPrev" to enable/disable one-direction swipes
+  * "cssWidthAndHeight" parameter could also accept "width" or "height" strings
+  * Better support for form elements on touch devices
+  * Lot of minor fixes
+
+## Swiper 2.6.1 - Updated on May 6, 2014
+
+  * Fix for fireCallback function when it is called externally
+
+## Swiper 2.6.0 - Updated on April 9, 2014
+
+  * New methods to enable/disable mousewheel control: enableMousewheeControl/disableMousewheeControl
+  * Minor fixes
+
+## Swiper 2.5.5 - Updated on March 23, 2014
+
+  * New API parameter: roundLengths
+  * Lot of minor fixes
 
 ## Swiper 2.5.0 - Updated on March 6, 2014